+++ profiler/src/test/BenchmarkBootupTestIntegration.java
@@ -1,5 +1,7 @@
-package grakn.benchmark.profiler;
+package grakn.benchmark.profiler.test;
 
+import grakn.benchmark.profiler.BootupException;
+import grakn.benchmark.profiler.GraknBenchmark;
 import grakn.benchmark.profiler.util.BenchmarkArguments;
 import org.junit.Rule;
 import org.junit.Test;
@@ -15,7 +17,7 @@ import java.nio.file.Paths;
  */
 public class BenchmarkBootupTestIntegration {
 
-    private final static Path WEB_CONTENT_CONFIG_PATH = Paths.get("profiler/test/resources/web_content/web_content_config_test.yml");
+    private final static Path WEB_CONTENT_CONFIG_PATH = Paths.get("test/resources/web_content/web_content_config_test.yml");
 
     @Rule
     public final ExpectedException expectedException = ExpectedException.none();
