+++ profiler/src/test/BenchmarkTestIntegration.java
@@ -1,5 +1,7 @@
-package grakn.benchmark.profiler;
+package grakn.benchmark.profiler.test;
 
+import grakn.benchmark.profiler.BootupException;
+import grakn.benchmark.profiler.GraknBenchmark;
 import grakn.benchmark.profiler.util.BenchmarkArguments;
 import grakn.core.GraknTxType;
 import grakn.core.Keyspace;
@@ -21,7 +23,7 @@ import java.util.UUID;
 import static grakn.core.graql.Graql.var;
 
 public class BenchmarkTestIntegration {
-    private final static Path WEB_CONTENT_CONFIG_PATH = Paths.get("profiler/test/resources/web_content/web_content_config_test.yml");
+    private final static Path WEB_CONTENT_CONFIG_PATH = Paths.get("test/resources/web_content/web_content_config_test.yml");
 
     @Rule
     public final ExpectedException expectedException = ExpectedException.none();
