+++ profiler/src/test/resources/BUILD
