+++ profiler/src/test/resources/web_content/queries.yml
