+++ profiler/src/test/resources/web_content/web_content_schema.gql
